terraform: fix inconsistent plan caused by changing config files
91af712e63af469dedb746c517d5b473845db642 attempted to resolve the issue
of Terraform reporting inconsistent plan, but it turns out it broke
updating the resource, as for Terraform, the changing value is fine'ish,
but new key appearing in the filed is not, hence we need to set the
fields to the empty value, if content of the config file is not known
(or empty). Then Terraform marks this fields as "Known After Apply".

Signed-off-by: Mateusz Gozdek <mgozdekof@gmail.com>